[PaaS] personalisation-service production (#61)
feat[cohorts]: introduce the concept of cohorts and implement its usage

cohorts are related to languages and defines language-locales in an unitary way or in a region way.
these defined cohorts are used to define regions and call for relevance score for calculating top
news scores.

it is being used to build the top_news_v1 and top_news_v2 requests. Also, on the v1/tabs request,
if a cohort cannot be defined for the call of top_news_v1 or v2, even when one of these tabs
are requested, the application will return a top_news_v0 feed to avoid users receiving tabs
that are empty due to lack of calculated cohort.

[PERSO-5439](https://onefootball.atlassian.net/browse/PERSO-5439)