[PaaS] personalisation-service production (#99)
feat[content application]: carousel sorting methodology is used to define sorting strategy on application

Introduced business logic to make sure that the stated sorting methodology behaviour is used in
order to set the sorting strategy on the aplication layer FOR CAROUSELS. The sorting strategy
can be sent from the application Find method call, and this is used for the main feed sorting
strategy pick. In the other hand, to decide the internal feeds (carousels) sorting strategy,
the one set in the sorting methodology behaviour is used to set it's strategy, and consequently,
the index adapter to use.

[PERSO-5433](https://onefootball.atlassian.net/browse/PERSO-5433)