[PaaS] creators-network production (#104)
chore(lint): enable `paralleltest` rules

Small note: the rule isn't flagging subtests, because it's not feasible at the moment, as too many tests would break with it. For now the linter only enforces the top level, but we should still be encouraged to use it also for sub-tests.

Small note 2: I had to refactor a couple of options in in rabbit.go because the way it exposed a global variable flagged a data race.