[PaaS] personalisation-service production (#166)
feat[hardcoded tabs]: introduce top_performing_videos_v1 feed

top_performing_videos_v1 introduces same behaviours as top_performing_videos_v0, with the difference
of the position. This one appears first in the list of items as an internal feed. This is necessary
in order to be used inside the top_news tabs that are requested from news API, thich should have
the carousel of top performing videos as the first item of the second page.

[PERSO-5433](https://onefootball.atlassian.net/browse/PERSO-5433)