[PaaS] creators-network production (#302)
feat(article-service): call clips api for jwplayer video details

As requested by Vasco, we need the `clips_id` whenever we have a jwplayer_video
content part. The suggested way to do so from Esterfano is to call the clips api.

Turns out that the clips API gives us everything, so we don't need to call the jwplayer api
and our db anymore.

The only thing that we're not confident in is the video url, as it returns a playlist m3u8
format rather than an mp4 link, which we previously had. We'll do some testing to see if that's okay.

Co-authored-by: Yasmina Garcia <yasmina.muro@onefootball.com>