[PaaS] creators-network production (#352)
test: simplify test language

As we discussed, the tests should start with the SUT under the right conditions for testing
rather than the test including building of the steps.

This commit moves all the steps into the dsl.NewSystemUnderTest() function and allows options
to be passed to setup the condition/state needed for the test.
Currently, only the code is moved inside the dsl layer to make the test less verbose. The next steps are
to modify the state in the database directly rather than invoking the API to do so.