[PaaS] creators-network production (#359)
fix: use int instead of int16 for bigger resolutions (#1723)

there were errors on the SLO for wordpress sync regarding parsing the
wordpress feed for a certain provider. It could potentially be because
the width of the media is higher than what an int16 (2^15) can store.
With this commit, we attempt to fix this issue if that actually is the
case.


![image](https://github.com/user-attachments/assets/089f960f-4d86-4627-9414-38c72af661f1)