[PaaS] creators-network production (#369)
chore: use the new `article-sourced-creators-network-yahoo` queue

Brenno introduced this lovely new `usecase` feature so now we can easily create
new queues of the same event with a different suffix.

Yahoo will be the guinea pig and once everything is confirmed to be working
we can migrate the article-translator and kill the separate app.toml there.