[PaaS] creators-network production (#372)
feat(authors): create endpoints to manage author names (#1714)

create two endpoints to get the author name (using the user from the
token)
and update the author name

this allows to implement the functionality on the content studio for the
authors to add their name

Notes about the implementation:
the endpoint requires the authors auth0 id to be passed as url
parameter, this parameter will be checked against the auth0 id in the
authorisation info. In case of a mismatch the request will be rejected
with status unauthorised. This is to leave space to future iterations if
we want to add a permission to let an admin edit other peoples author
name

Co-authored-by: Lukas Spiss <lukas.spiss@onefootball.com>