[PaaS] article-intelligence production (#379)
feat(translations): add langfuse prompt manager (#83)

missing

- [x] ~~mock Langfuse() globally to stop the test execution from hanging
after success~~ We solved this, by killing all hanging threads at the
end
- [ ] refactor Evals

## the problem with Evals:

In the original implementation, a candidate was responsible of
translating an article. In fact, we had a prompt that received an entire
article and was asked to translate title and content together with the
same LLM.
The Evals were based on this logic, and they were running on a list of
candidates translating articles.

Now the logic changed, and a candidate is responsible of translating
text. There are two different candidates used to translate an article
(this means two different providers, two different models and two
different prompts, by random chance only provider and model can be the
same, the prompt will always be different).

So the Eval logic needs to be re-thinked. How do we run evals?

---------

Co-authored-by: Gabriele Boscaratto <gabriele.boscaratto@onefootball.com>