[PaaS] funtik production (#475)
fix: skip notification for draft articles

We noticed that draft articles were sending article-sourced events, but as a hack
they were sent without a language. This is very awkward for consumers, especially
because the schema advertises the property as required.

We already skip the notification for drafts when the article comes from Content Studio,
as we handle the event directly there, but for 3rd party articles we still rely on
Funtik.

Co-authored-by: Yasmina Garcia <yasmina.muro@onefootball.com>
Co-authored-by: Minhaj Khan <minhaj.khan@onefootball.com>