[PaaS] personalisation-service production (#489)
chore[observability]: do not set code message as status message

avoid redundant error message sharing in the status message sent. The error is already there on
the recordError method and can be found in it's specific exception. Seting the error message in
the status message can lead to misleading investigations, since the error on the main span
'exception.message' can, in somecases, not be the reason for the 'status code error' from the
span. Better to always check every exception span given to a trace to come to this conclusion.

[PERSO-5477](https://onefootball.atlassian.net/browse/PERSO-5477)