[PaaS] creators-network production (#492)
chore: replace public link with the `source_url` from articles table

Ref: https://onefootball.slack.com/archives/GKAT0DG11/p1743493180166829

We made a wrong assumption that the link prop is basically always the onefootball direct link.

However, the link is actually meant to be the canonical link. So instead we should just trust the
`source_url`, which should be our internal link (for homegrown articles) and the 3rd party link for
3rd party articles.

We still need to set and backfill the `source_url` for our articles, but it's not that urgent as
the web generates it themselves.

Co-authored-by: Yasmina Garcia <yasmina.muro@onefootball.com>
Co-authored-by: Minhaj Khan <minhaj.khan@onefootball.com>
Co-authored-by: Mikjail Salazar <mikjail.salazar@onefootball.com>
Co-authored-by: Gauraang Bhartiya <gauraang.bhartiya@onefootball.com>
Co-authored-by: Chunfeng Xia <chunfeng.xia@onefootball.com>