[PaaS] creators-network production (#546)
feat: prepare tags to save them in the database (SRC-3520) (#1733)

This commit **prepares** the tags so that they can be stored in the
database, and eventually be served to the fans.

To serve the article _with tags_, they are required to be present in our
database. Since the tags come from content-classifier, we need to plug
into the article-classified event in order to get the tags and store
them. However, the event only contains identifiers of the tags; we need
to prepare the tags first to display them to the fans.