[PaaS] creators-network production (#675)
refactor: set up system under test by setting it in the database (#1739)

Context:
The initial state of the tests was a bit verbose; the tests generated
the state first and then tested the functionality. After a discussion,
we decided that the test should only test the functionality, and the
system under test should be prepared beforehand.

As a quick fix, we moved talking to the system into the constructor to
remove verbosity from the tests, however, the tests were still invoking
REST endpoints to the application to get the system into the right
state.

This commit allows the system under test to be in the state by just
putting the right values in the database, instead of invoking a sequence
of steps.