[PaaS] creators-network production (#713)
refactor: use author_id from articles table instead of fetching from article revisions (#1740)

Since the author_id was added later in the articles table, we do not
need to fetch the oldest revision to know the author.
This author is added to the article table as soon as the article is
created, which equates to the first/oldest revision created.

relates to #1659