[PaaS] personalisation-service production (#734)
test[relevance adapter]: update unit test to new implementation of parser from pgx

pgx introduced a new logic to try to parse types from the .Scan method, which is good, but broke
the expectation from unit tests on this implementation. Tests were updated to purposly send an non
parseable type, so the unit test would be triggered for error on scan parse.