[PaaS] personalisation-service production (#740)
feat[hardcoded tabs]: introduce tabs 'Callbacks'

Callbacks are a slice of tabs definitions that a tab can have. When used, callbacks are called when
the main feed contents result where lesser then the minimum desired. This minimum desired value
is controlled by the application who uses the tab.

This is a first step to move tab callback definition to tab domain. Further, the application layer
has to be able to interpret this slice and make the callback calls if necessary.

[PERSO-5520](https://onefootball.atlassian.net/browse/PERSO-5520)