[PaaS] creators-network production (#767)
fix: roll-back provider creation on auth0 error

Ref: https://onefootball.slack.com/archives/GKAT0DG11/p1746636904732939

Previously the code didn't make use of a transaction, but instead only relied
on manually deleting the inserted provider, whenever there was an error with
Funtik.

This mean that every time there was an error with Auth0 (usually a naming issue like
having a comma or something else "illegal"), we ended up keeping the inserted
rows, resulting in many duplicate unusable providers.

Instead the code is now wrapped in a transaction so regardless of what part fails,
the insert can be rolled back easily.