[PaaS] creators-network production (#1190)
feat: enable translations for wordpress provider America y ya (SRC-3576)

This commit enables creating translations in draft mode only for the
specified provider America y ya. Currently, the translations will only
be done from Spanish to English in a draft state.

Please note that the language configuration for auto-publishing is not
configured by provider ID at the moment. If we enable auto-publishing
for a provider, it will auto-publish for all the languages we have
enabled to be auto-published previously for original content
(OneFootball articles).


---------

Co-authored-by: Chunfeng Xia <chunfeng.xia@onefootball.com>